bin folder, forgot to commit vect2.c with last commit. Engine.c now displays how points are in the plane

+++ src/vect2.c
@@ -11,15 +11,15 @@ vect2 vect2_diff(vect2 v1,vect2 v2){
 /*
 Divide vector coordinates by a floating point number
 */
-vect2 divide(vect2 v,float n){
+vect2 divide(vect2 v,double n){
     vect2 temp = {.x = v.x / n,.y = v.y / n};
     return temp;
 }
 
-float vect2_length(vect2 v){
+double vect2_length(vect2 v){
     return sqrtf(v.x * v.x + v.y * v.y);
 }
 
-float dot_product(vect2 v1,vect2 v2){
+double dot_product(vect2 v1,vect2 v2){
     return v1.x * v2.x + v1.y * v2.y;
 }
